st/texture-cache: Switch to integers for resource scales
Since clutter_actor_get_resource_scale() now returns an unsigned integer
instead of a float for the scale, switch to ints for storing that scale
in the texture cache, too.

https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1287

+++ src/st/st-icon.c
@@ -409,7 +409,7 @@ st_icon_update (StIcon *icon)
   gint paint_scale;
   ClutterActor *stage;
   StThemeContext *context;
-  float resource_scale;
+  unsigned int resource_scale;
 
   if (priv->pending_texture)
     {
